Corrected type of -affectedRows to match mysql's internal type

+++ Source/MySQL/CSMySQLPreparedStatement.h
@@ -26,6 +26,6 @@
     id paramBinds;
 }
 
-- (int)affectedRows;
+- (unsigned long long)affectedRows;
 
 @end
